LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-spacing/tentative/parsing/text-spacing-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 4: parsing text-autospace with valid values</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value('text-spacing', 'auto none');
test_invalid_value('text-spacing', 'none auto');
test_invalid_value('text-spacing', 'none garbage');
test_invalid_value('text-spacing', 'auto garbage');
test_invalid_value('text-spacing', 'garbage auto');
test_invalid_value('text-spacing', 'garbage none');
</script>
</body>
</html>
